Solve the number of primes in a diagonal (P058)
Supposed a square of side 7 that look like this:

37 36 35 34 33 32 31
38 17 16 15 14 13 30
39 18  5  4  3 12 29
40 19  6  1  2 11 28
41 20  7  8  9 10 27
42 21 22 23 24 25 26
43 44 45 46 47 48 49

One will find out that the ratio of number of primes in the diagonals to
the number of items lying on the diagonals is approximately 62% (8/13).
The program will find out what is the length of the side which will have a
ratio of less than 10% in its diagonal.