Find the sum of the ascii equivalent P059
Display the decode text when the text has 'the' and 'and'. Those two words may
be in the text and are common in the English language. There are better ways to
solve it though rather than prompting the user if the decoded text look right
and if not find other combination.