Determine the sum of cyclic pol numbers (P061)
Find the sum of six polynomial numbers each of which is a triangle,
square, pentagonal, hexagonal, heptagonal and octagonal numbers.

The solution finds all the polynomial numbers within the range and put it
in a dictonary. The dictionary has the number of sides as a key and another
dictonary for the number generated by the side (triangular, square, etc.). It
is implemented as dictonary so that the head and the tail of the number can be
access easily.

Using recursion makes the implementation more understandable because it is the
nature of the problem.

Because of the way traverse is implemented, duplicate sequence are also
generated.