Find the smallest cube of which permutation are cube (P062)
Find the smallest of which 5 of its permutations are also cube.

To make things easier, generate the cube first and then sort its digit.
In doing this, some cubes which is a permuation of the other will match.
Because the cube is generated from lowest to highest, store the number
in the value corresponding the the sorted digit if it does not exist. This
will ensure that the smallest number will be printed if ever needed. If it does
exist, increment the counter for that. Thus the smallest number that satisfy
the problem solution will be printed.