Find the largest x in x^2 - D y^2 = 1 (P066)
For all the minimal values of x when D <= 1000, find the largest
value of x.

This is a Pell's equation. The solution is to find a convergent, h/k
in the continued fraction of the square root of N, then compare the
values to each other to find the largest and print the value of D which
satisfy the problem.